fix: search reducer
